meta(): restructures project to avoid cyclic dependencies
