--- mockserver/api/urls.py
@@ -1,3 +1,0 @@
-from mocks.urls import router
-
-urlpatterns = router.urls
