+++ mockserver/authentication/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
 
 
-class AuthConfig(AppConfig):
-    name = 'authentication'
+class ApiConfig(AppConfig):
+    name = 'api'
