+++ mockserver/mocks/urls.py
@@ -1,7 +1,6 @@
-from django.urls import re_path
 from rest_framework import routers
 
-from mocks.views import fetch_mock, MockViewset, HeaderTypeViewset, HttpVerbViewset
+from mocks.views import MockViewset, HeaderTypeViewset, HttpVerbViewset
 
 
 router = routers.SimpleRouter()
@@ -10,6 +9,4 @@ router.register('header-types', HeaderTypeViewset)
 router.register('http-verbs', HttpVerbViewset)
 
 
-urlpatterns = [
-    re_path(r"^.*/$", fetch_mock)
-]
+urlpatterns = router.urls
