+++ mockserver/mockserver/urls.py
@@ -19,6 +19,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/v1/', include('api.urls')),
-    path('', include('mocks.urls')),
+    path('api/v1/', include('mocks.urls')),
+    path('', include('tenants.urls')),
 ]
