+++ mockserver/tenants/urls.py
@@ -1,0 +1,8 @@
+from django.urls import re_path
+
+from tenants.views import fetch_mock
+
+
+urlpatterns = [
+    re_path(r"^.*/$", fetch_mock)
+]
